ocr_service/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1
pytest==8.3.4
opencv-python-headless
easyocr
Pillow
rapidfuzz
imagehash
